Restructured 'renderer' package.
