+++ game-engine/src/main/java/gameengine/engine/renderer/IRenderPass.java
@@ -1,5 +1,7 @@
 package gameengine.engine.renderer;
 
+import gameengine.engine.renderer.uniform.UniformManager;
+
 public interface IRenderPass {
 	public UniformManager getUniformManager();
 }
